+++ avs_server/src/lib.rs
@@ -18,3 +18,5 @@ pub mod ca;
 pub mod config;
 pub mod csr;
 pub mod server;
+
+pub use policies;