+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.6.0 (2026-07-07)
+
+
+### Features
+
+* add DEVELOPMENT_CB_CERTIFICATE support
+* add mTLS and TLS development policy hints
+* integrate devkit/check_checksums --update
+
 ## 0.5.0 (2026-07-02)